Moved handler functions to another file, added http error responses
