+++ oss-cloud-backend/handler.js
@@ -1,14 +1,6 @@
 "use strict";
 
-const access_token = process.env.GITHUB_ACCESS_TOKEN;
-const Octokit = require('@octokit/rest') // github api library
-const octokit = new Octokit({
-  auth: access_token,
-})
-
-octokit.hook.error('request', async (error, options) => {
-  throw error
-})
+var getPullsHandler = require('./getPullsHandler.js')
 
 module.exports.hello = async event => {
   return {
@@ -32,75 +24,19 @@ module.exports.hello = async event => {
 // }
 module.exports.getPulls = (event, context, callback) => {
   var body = JSON.parse(event.body);
-  getForkedRepos(body.username) // get all repos of a user
-    .then(repos => getRepoDetails(repos))
-    .then(repos => getParentRepos(repos))
-    .then(repos => getUserPulls(body.username, repos))
+  getPullsHandler.getForkedRepos(body.username) // get all repos of a user
+    .then(repos => getPullsHandler.getRepoDetails(repos))
+    .then(repos => getPullsHandler.getParentRepos(repos))
+    .then(repos => getPullsHandler.getUserPulls(body.username, repos))
     .then(pulls => callback(null, pulls))
-    .catch((err) => callback(err, {status: 500, body: err.message}))
+    .catch(err => {
+      let response = {
+        statusCode: err.status,
+        body: JSON.stringify({
+          message: err.message
+        })
+      }
+      callback(null, response);
+    })
 };
 
-
-// returns a single repo
-var getRepo = async (owner, repo) => {
-  return octokit.repos.get({
-    owner: owner,
-    repo: repo,
-  }).then(({data, headers, status}) => {
-    console.log("repo: ", data.name, " status", status)
-    return data;
-  })
-}
-
-var getForkedRepos = async(username) => {
-  return octokit.search.repos({
-    q: "user:" + username + "+fork:only"
-  }).then(({data, headers, status}) => {
-    return data.items;
-  })
-}
-
-
-// makes a call to github api for each repo in array, returning a more detailed representation of the repo
-var getRepoDetails = async (repos) => {
-  const repoPromises = repos.map(async (repo) => {
-    return getRepo(repo.owner.login, repo.name)
-  })
-
-  return await Promise.all(repoPromises);
-}
-
-// returns the parent repos from an array of forked repos (details needed)
-var getParentRepos = async (repos) => {
-  const repoPromises = repos.map(async (repo) => {
-    return getRepo(repo.parent.owner.login, repo.parent.name);
-  })
-
-  return await Promise.all(repoPromises);
-
-}
-
-// retrieves pulls for given user for each repo in given array
-var getUserPulls = async (username, repos) => {
-  let pulls = [];
-  const pullsPromises = repos.map(async (repo) => {
-    let repoPulls = await searchUserPulls(username, repo);
-    pulls = pulls.concat(repoPulls);
-    return repoPulls;
-  })
-
-  await Promise.all(pullsPromises);
-
-  return pulls;
-}
-
-// returns pulls from a repo where given user is the author
-var searchUserPulls = async(username, repo) => {
-  console.log("searching pull requests: ", username, repo.name)
-  return octokit.search.issuesAndPullRequests({
-    q: "repo:" + repo.owner.login + "/" + repo.name + "+author:" + username + "+is:pr",
-    per_page: 100
-  }).then(({data, headers, status}) => {
-    return data.items;
-  })
-}
